moved some things from twaliProvider to Profile
